// lib/machine/interfaces.d.ts
export interface MachineArguments extends Array<InitDirective | StateDirective | ParallelDirective> {
}
export interface InitDirective {
    initial?: InitialDirective;
    context?: ContextDirective;
    freeze?: ShouldFreezeDirective;
    history?: HistoryDirective;
}
export interface Context {
    [key: string]: any;
}
export interface TransitionDirective {
    transition: string;
    target: string;
    guards: GuardsDirective;
    exit?: ExitDirective[];
}
export interface TransitionsDirective {
    [key: string]: TransitionDirective;
}
export interface ImmediateDirective {
    immediate: string;
    guards: GuardsDirective;
}
export interface NestedImmediateDirective extends ImmediateDirective {
    immediate: string;
    guards: GuardsDirective;
}
export interface ParallelImmediateDirective extends ImmediateDirective {
    immediate: string;
    guards: GuardsDirective;
}
export interface Pulse {
    (context: Context, payload?: any): Promise<void | any> | void;
}
export interface PulseDirective {
    pulse: Pulse;
    success?: string | PulseDirective;
    failure?: string | PulseDirective;
    transition?: string;
}
export interface ExitDirective {
    pulse: Pulse;
    failure?: string;
}
export interface Guard {
    (context: Context, payload?: any): boolean | any;
}
export interface GuardDirective {
    guard: Guard;
    failure?: string | PulseDirective;
}
export interface NestedGuardDirective extends GuardDirective {
    machine: Machine;
    failure?: string | PulseDirective;
}
export interface GuardsDirective extends Array<GuardDirective | NestedGuardDirective> {
}
export interface DescriptionDirective {
    description: string;
}
export interface RunCollection extends Array<NestedMachineDirective | PulseDirective | TransitionDirective | ImmediateDirective | DescriptionDirective> {
}
export interface StateDirective {
    name: string;
    run: PulseDirective[];
    on: TransitionsDirective;
    immediate: ImmediateDirective[];
    args: (NestedMachineDirective | PulseDirective | TransitionDirective | ImmediateDirective | DescriptionDirective)[];
    type: string;
    nested: NestedMachineDirective[];
    description?: string;
}
export interface InfoStateDirective extends StateDirective {
    type: "info";
}
export interface WarningStateDirective extends StateDirective {
    type: "warning";
}
export interface DangerStateDirective extends StateDirective {
    type: "danger";
}
export interface PrimaryStateDirective extends StateDirective {
    type: "primary";
}
export interface SuccessStateDirective extends StateDirective {
    type: "success";
}
export interface StatesDirective {
    [key: string]: StateDirective;
}
export interface ParallelDirective {
    parallel: Record<string, Machine>;
}
export interface Machine {
    id: string;
    title: string | null;
    context: Context;
    isAsync: boolean;
    states: StatesDirective;
    initial: string;
    current: string;
    frozen: boolean;
    fatal?: Error;
    history: string[];
    historyLimit?: number;
    parallel: ParallelDirective["parallel"];
}
export interface NestedMachineDirective {
    machine: Machine;
    transition?: string;
}
export interface NestedMachineWithTransitionDirective extends NestedMachineDirective {
    transition: string;
}
export interface ContextDirective {
    context: Context;
}
export interface InitialDirective {
    initial: string;
}
export interface ShouldFreezeDirective {
    freeze: boolean;
}
export interface HistoryDirective {
    history: number;
}
export declare enum HistoryType {
    Transition = "Transition",
    Pulse = "Pulse",
    AsyncPulse = "Async Pulse",
    State = "State",
    Guard = "Guard"
}
export declare const START_EVENT = "__start__";
//# sourceMappingURL=interfaces.d.ts.map